Release 20240822.1
Add Footer

+++ net-react.client/src/DeleteCategory.tsx
@@ -1,15 +1,22 @@
-import React, { useState } from 'react';
+import { useState, FC } from 'react';
 import axios from 'axios';
 import { DotNetApi } from './helpers/DotNetApi';
 
-const DeleteCategory: React.FC = () => {
-    const [id, setId] = useState('');
+const DeleteCategory: FC = () => {
+    const [id, setId] = useState<number | undefined>(undefined);
 
-    const handleDeleteCategory = () => {
+    const handleDeleteCategory = async (event: React.FormEvent<HTMLFormElement>) => {
+        event.preventDefault();
+        if (id === undefined) {
+            alert("Category ID is required");
+            return;
+        }
+        //axios.delete(DotNetApi + 'Category/Delete/' + id)
         axios.delete(DotNetApi + `Category/Delete/${id}`)
             .then(response => {
                 console.log('Category deleted: ', response.data);
                 alert("Deleted. Reload page");
+                location.reload();
             })
             .catch(error => {
                 console.error('There was an error!', error);
@@ -18,15 +25,19 @@ const DeleteCategory: React.FC = () => {
     };
 
     return (
-        <div>
-            <h2>Delete Category</h2>
-            <input
-                type="text"
-                value={id}
-                onChange={(e) => setId(e.target.value)}
-                placeholder="Category ID"
-            />
-            <button onClick={handleDeleteCategory}>Delete</button>
+        <div className="container">
+            <form onSubmit={handleDeleteCategory}>
+                <div>
+                    <h2>Delete Category</h2>
+                    <input
+                        type="text"
+                        value={id}
+                        onChange={(e) => setId(e.target.value !== '' ? parseInt(e.target.value) : undefined)}
+                        placeholder="Category ID"
+                    />
+                    <button type="submit">Delete</button>
+                </div>
+            </form>
         </div>
     );
 };
